TPI-341 удалил лишний компонент
