+++ src/components/form/Checkbox/Checkbox.types.ts
@@ -1,25 +1,11 @@
-import React, {ChangeEventHandler} from 'react';
+import React from 'react';
 import {FieldHelperProps} from 'formik/dist/types';
 
 export type CheckboxState = 'on' | 'off' | 'partial';
 
-export type ICheckboxProps = IRoundedCheckboxProps | IDefaultCheckboxProps;
-
-export interface IRoundedCheckboxProps extends React.InputHTMLAttributes<HTMLInputElement> {
-    rounded?: true;
-    checked?: boolean;
-    name?: string;
-    value: string;
-    disabled?: boolean;
-    className?: string;
-    error?: string;
-    onChange?: ChangeEventHandler<HTMLInputElement>;
-    onBlur?: (event: React.FormEvent<HTMLInputElement>) => void;
-}
-
-export interface IDefaultCheckboxProps
+export interface ICheckboxProps
     extends React.DetailedHTMLProps<React.InputHTMLAttributes<HTMLInputElement>, HTMLInputElement> {
-    rounded?: false;
+    rounded?: boolean;
     text?: string;
     caption?: string;
     small?: boolean;
@@ -28,7 +14,7 @@ export interface IDefaultCheckboxProps
 }
 
 // checkbox props with formik integrations
-export type CheckboxWithFormikType = IDefaultCheckboxProps & Partial<FieldHelperProps<boolean>>;
+export type CheckboxWithFormikType = ICheckboxProps & Partial<FieldHelperProps<boolean>>;
 
 export interface IThreeStateCheckboxProps extends CheckboxWithFormikType {
     value: CheckboxState;
