+++ src/components/form/Checkbox/CheckboxField.tsx
@@ -1,9 +1,9 @@
 import React, {FC} from 'react';
 import {Checkbox} from './index';
 import {FieldHookConfig, useField} from 'formik';
-import {IDefaultCheckboxProps} from './Checkbox.types';
+import {ICheckboxProps} from './Checkbox.types';
 
-export const CheckboxField: FC<IDefaultCheckboxProps & FieldHookConfig<boolean>> = (props) => {
+export const CheckboxField: FC<ICheckboxProps & FieldHookConfig<boolean>> = (props) => {
     const {name, validate, type, multiple, value} = props;
     const [field, meta, helpers] = useField({name, validate, type, multiple, value});
     const tmpChecked = Array.isArray(field.value) ? field.value.includes(props.value) : props.checked || field.value;
