--- src/components/form/Checkbox/cform__checkbox--rounded.less
@@ -1,120 +1,0 @@
-@import "../../../assets/css/bundle.less";
-.cform__checkbox--rounded {
-  --background-color: initial;
-  --border-color: @c-dark-gray;
-  --icon-color: initial;
-
-  display: inline-block;
-  position: relative;
-  width: 32px;
-  height: 32px;
-  border-radius: 50%;
-  user-select: none;
-
-  @media(max-width: @rsp-mobile-max) {
-    width: 40px;
-    height: 40px;
-  }
-
-  input {
-    position: absolute;
-    width: 0;
-    height: 0;
-    opacity: 0;
-  }
-
-  input:disabled + label {
-    --icon-color: @c-gray-35;
-    --border-color: transparent;
-    --background-color: @c-light-gray;
-    cursor: auto;
-  }
-
-  input:not(:checked):disabled + label {
-    --icon-color: @c-gray-35;
-    --border-color: @c-dark-gray-fade-35;
-    --background-color: transparent;
-  }
-
-  input:not(:disabled) {
-    + label {
-      .on-hover(active-fallback; {
-        box-shadow: inset 0 0 0 2px var(--border-color);
-      });
-    }
-
-    &:checked + label {
-      --icon-color: @c-white;
-      --border-color: transparent;
-      --background-color: @c-primary-blue;
-    }
-  }
-
-  input:checked + label &__icon {
-    display: block;
-  }
-
-  label {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    position: absolute;
-    top: 0;
-    right: 0;
-    bottom: 0;
-    left: 0;
-    border-radius: 50%;
-    background-color: var(--background-color);
-    box-shadow: inset 0 0 0 1px var(--border-color);
-    transition: all @duration-blink;
-    cursor: pointer;
-
-    svg.cicon * {
-      fill: var(--icon-color);
-    }
-  }
-
-  &__icon {
-    display: none;
-  }
-
-  &__text {
-    white-space: nowrap;
-    color: var(--icon-color);
-    font-family: @ff-system;
-    font-size: 14px;
-    line-height: 22px;
-    letter-spacing: 0.01px;
-
-    @media(max-width: @rsp-mobile-max) {
-      font-size: 13px;
-      line-height: 20px;
-    }
-
-  }
-
-  .mtheme--lightpic & {
-    --icon-color: @c-black;
-    --border-color: @c-gray-50-fade-35;
-
-    input:checked:disabled + label {
-      --background-color: @c-gray-50-fade-20;
-    }
-  }
-
-  .mtheme--darkpic & {
-    --icon-color: @c-white;
-    --border-color: @c-gray-50-fade-35;
-
-    input:disabled + label {
-      --icon-color: @c-gray-75;
-      --border-color: @c-gray-50-fade-35;
-    }
-
-    input:checked:disabled + label {
-      --icon-color: @c-gray-75;
-      --background-color: @c-gray-50-fade-20;
-      box-shadow: none !important;
-    }
-  }
-}
