+++ src/components/form/Checkbox/cform__checkbox.less
@@ -70,6 +70,16 @@
         opacity: 1;
       }
     }
+
+    &--rounded {
+      width: 32px;
+      height: 32px;
+      border-radius: 50%;
+      svg {
+        top: 0;
+        left: 0;
+      }
+    }
   }
 
   &__body {
