+++ src/components/form/Checkbox/index.ts
@@ -1,4 +1,4 @@
-import {Checkbox} from './Checkbox';
 import {CheckboxField} from './CheckboxField';
+import {Checkbox} from './Checkbox';
 
 export {CheckboxField, Checkbox};
